pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-ai"
version = "0.3.0"
description = "NEXUS AI - Neural EXecution and Understanding System"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "NEXUS AI Team"},
]
keywords = ["ai", "repl", "claude", "terminal", "interactive"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Tools",
    "Topic :: Terminals",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.3.0",
    "prompt_toolkit>=3.0.0",
    "pygments>=2.15.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]
gemini-api = [
    "google-generativeai>=0.3.0",
]
all = [
    "google-generativeai>=0.3.0",
]

[project.scripts]
nexus-ai = "nexus_ai.main:main"
nexus = "nexus_ai.main:main"

[project.urls]
Homepage = "https://github.com/your-org/nexus-ai"
Repository = "https://github.com/your-org/nexus-ai"
Documentation = "https://github.com/your-org/nexus-ai#readme"
"Bug Reports" = "https://github.com/your-org/nexus-ai/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["nexus_ai*"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312", "py313"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true